requirements.txt
JPype1==1.4.1
JPype1==1.4.1
pygame==2.1.3